feat(wake): Add not applicable to wake calculator display (#426)
It was sometimes hard to tell when there wasn't a wake interval because
it couldn't be worked out (or because of a bug) and when there legitimately isn't one.
This makes the default display to be a double dash. If two aircraft are selected such that
no interval applies (e.g. LM followed by LM), it will instead display N/A which makes it clear
to the user that there is no applicable interval.

Resolves #425