fix: oceanic tag items crash
Fixes a crash if you try to open the oceanic clearance dialog when using
the oceanic clearance tag items. This crash happens because the mutex covering
the clearance map is locked at the start of TagFunction and wont unlock until
the window has been closed. This causes the same ES thread to come along later (to
display the tag items, try to lock the mutex, and everything explodes.

This change fixes the issue by scoping the mutex to a new function that has only
the responsibility to grab the value from the map (or use the default if required). That
way, subsequent calls can access the map properly.

Fixes #511