perf: predetermine history trail draw function
On every render, we were checking the fields to determine the type of dot
we are rendering. However, we don't need to do this, as the dot type can only change
as part of an ASR load or configuration menu change.

This change saves the draw function as a lambda when something changes, allowing us to
reduce the render time by approximately 25% - went from 12ms to 8ms in testing.